feat: rename package to zensical-macros-utils and bump version to 0.10.0 (#202)
- Rename mkdocs_macros_utils/ -> zensical_macros_utils/
- Rename tests/python/mkdocs_macros_utils/ -> tests/python/zensical_macros_utils/
- Remove DeprecationWarning from __init__.py (was for tombstone release)
- Update pyproject.toml: name, version 0.0.11 -> 0.10.0, URLs, packages, mypy files
- Update zensical.toml: site_name, site_url, repo_name, repo_url, modules
- Update all test imports and mocker.patch paths
- Update CI workflows: test_multi_os.yml path filters, update_version_and_release.yml release name
- Update README.md, README_PyPI.md, docs/index.md, CLAUDE.md
- Update package.json, package-lock.json: name, version 0.0.11 -> 0.10.0
- Update .vscode/settings.json
- Fix CLAUDE.md markdownlint errors (MD032, MD060)
