+++ .github/workflows/test_multi_os.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - 'main'
     paths:
-      - 'mkdocs_macros_utils/**'
+      - 'zensical_macros_utils/**'
       - 'tests/**'
       - 'pyproject.toml'
       - 'uv.lock'
@@ -44,7 +44,7 @@ jobs:
       - name: Check file existence
         id: check_file
         run: |
-          if [ -f "mkdocs_macros_utils/staking/ton_whales_staking_dashboard.py" ]; then
+          if [ -f "zensical_macros_utils/staking/ton_whales_staking_dashboard.py" ]; then
             echo "file_exists=true" >> "$GITHUB_OUTPUT"
           else
             echo "file_exists=false" >> "$GITHUB_OUTPUT"
@@ -170,7 +170,7 @@ jobs:
 
             # テスト対象外のファイルに関する注記を追加（ファイルの存在確認結果に基づく）
             commit_hash=${GITHUB_SHA}
-            file_path_1="mkdocs_macros_utils/staking/ton_whales_staking_dashboard.py"
+            file_path_1="zensical_macros_utils/staking/ton_whales_staking_dashboard.py"
             file_path_2="staking/ton_whales_staking_dashboard.py"
             file_link="https://github.com/${GITHUB_REPOSITORY}/blob/${commit_hash}/${file_path_2}"
 
@@ -205,8 +205,8 @@ jobs:
           # カバレッジレポート内のリンクを有効にするためにファイルパスを修正する
           sed -i '
             s|/blob/undefined/\([^"]*\)|/blob/'"${commit_hash}"'/\1|g;
-            s|/blob/\([a-f0-9]*\)/\([^"]*\)|/blob/\1/mkdocs_macros_utils/\2|g;
-            s|/blob/\([a-f0-9]*\)/mkdocs_macros_utils/README\.md|/blob/\1/README.md|g
+            s|/blob/\([a-f0-9]*\)/\([^"]*\)|/blob/\1/zensical_macros_utils/\2|g;
+            s|/blob/\([a-f0-9]*\)/zensical_macros_utils/README\.md|/blob/\1/README.md|g
           ' README.md
       - name: Commit and push
         run: |
