+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "python.testing.pytestArgs": [
-        "mkdocs_macros_utils"
+        "zensical_macros_utils"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true
