+++ README_PyPI.md
@@ -1,11 +1,8 @@
-# mkdocs-macros-utils
+# zensical-macros-utils
 
-> **Deprecated:** This package is no longer maintained.
-> Please migrate to **[zensical-macros-utils](https://pypi.org/project/zensical-macros-utils/)** instead.
+[zensical-macros-utils](https://7rikazhexde.github.io/zensical-macros-utils/) is a [zensical](https://zensical.org/)-based project that provides macros to extend cards, code blocks, etc, in MkDocs documents.
 
-[mkdocs-macros-utils](https://7rikazhexde.github.io/mkdocs-macros-utils/) is a [zensical](https://zensical.org/)-based project that provides macros to extend cards, code blocks, etc, in MkDocs documents.
-
-[![pages-build-deployment](https://github.com/7rikazhexde/mkdocs-macros-utils/actions/workflows/pages/pages-build-deployment/badge.svg?branch=gh-pages)](https://github.com/7rikazhexde/mkdocs-macros-utils/actions/workflows/pages/pages-build-deployment) [![DOCS](https://img.shields.io/badge/Docs-Click%20Here-blue?colorA=24292e&colorB=0366d6&logo=github)](https://7rikazhexde.github.io/mkdocs-macros-utils/)
+[![pages-build-deployment](https://github.com/7rikazhexde/zensical-macros-utils/actions/workflows/pages/pages-build-deployment/badge.svg?branch=gh-pages)](https://github.com/7rikazhexde/zensical-macros-utils/actions/workflows/pages/pages-build-deployment) [![DOCS](https://img.shields.io/badge/Docs-Click%20Here-blue?colorA=24292e&colorB=0366d6&logo=github)](https://7rikazhexde.github.io/zensical-macros-utils/)
 
 ## Features
 
@@ -15,38 +12,38 @@
 
 ## Usage
 
-### Install `mkdocs-macros-utils`
+### Install `zensical-macros-utils`
 
 ```bash
 # For pip
-pip install mkdocs-macros-utils
+pip install zensical-macros-utils
 
 # For uv
-uv add mkdocs-macros-utils
+uv add zensical-macros-utils
 ```
 
 ### Config settings
 
-1. Add the extension to your `mkdocs.yml`
+1. Add the extension to your `zensical.toml`
 
-    ```yaml
-    markdown_extensions:
-      - zensical.extensions.macros:
-          modules: [mkdocs_macros_utils]
+    ```toml
+    extra_css = [
+        "stylesheets/macros-utils/link-card.css",
+        "stylesheets/macros-utils/gist-cb.css",
+        "stylesheets/macros-utils/x-twitter-link-card.css",
+    ]
 
-    extra:
-      debug:
-        link_card: false  # Set to true for debug logging
-        gist_codeblock: false
-        x_twitter_card: false
+    extra_javascript = [
+        "javascripts/macros-utils/x-twitter-widget.js",
+    ]
 
-    extra_css:
-      - stylesheets/macros-utils/link-card.css
-      - stylesheets/macros-utils/gist-cb.css
-      - stylesheets/macros-utils/x-twitter-link-card.css
+    [project.plugins.macros]
+    modules = ["zensical_macros_utils"]
 
-    extra_javascript:
-      - javascripts/macros-utils/x-twitter-widget.js
+    [project.extra.debug]
+    link_card = false
+    gist_codeblock = false
+    x_twitter_card = false
     ```
 
 1. Start the development server
@@ -59,7 +56,7 @@ The plugin will automatically create the required directories and copy CSS/JS fi
 
 ## Documentation
 
-For detailed usage and examples, please see the [documentation](https://7rikazhexde.github.io/mkdocs-macros-utils/).
+For detailed usage and examples, please see the [documentation](https://7rikazhexde.github.io/zensical-macros-utils/).
 
 ## License
