+++ zensical_macros_utils/__init__.py
@@ -5,7 +5,6 @@ Zensical macros module for enhanced documentation components.
 import logging
 import os
 import shutil
-import warnings
 from pathlib import Path
 from typing import Any, Dict
 
@@ -14,14 +13,6 @@ from zensical.extensions.macros import MacroEnv
 from . import gist_codeblock
 from . import link_card
 from . import x_twitter_card
-
-warnings.warn(
-    "mkdocs-macros-utils is deprecated and will no longer be maintained. "
-    "Please migrate to zensical-macros-utils: "
-    "https://pypi.org/project/zensical-macros-utils/",
-    DeprecationWarning,
-    stacklevel=2,
-)
 
 logger = logging.getLogger("zensical.macros-utils")
